added ASP.NET Identity

+++ Data/Models/Message.cs
@@ -1,14 +1,13 @@
 using System;
 using System.ComponentModel.DataAnnotations.Schema;
-using Topicly.Data;
 
 namespace Data.Models
 {
     public class Message
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         
-        public int ChatId { get; set; }
+        public long ChatId { get; set; }
         [ForeignKey(nameof(ChatId))]
         public Chat Chat { get; set; }
